Version 2.0.0 (#2363)

+++ tests/errors/semantic/blocking/INSERT_LIST_INCOMPATIBLE_TYPES_0.py
@@ -3,6 +3,6 @@
 import numpy as np
 
 a = [1,2,3]
-b = np.int32(4)
+b = np.int16(4)
 a.insert(3, b)
